pyproject.toml
[project]
name = "agent-sandbox"
version = "0.1.0"
description = "agent-sandbox"
authors = [
    {name = "imtlib", email = "yuya78858845@gmail.com"},
]
dependencies = [
    "langchain-aws>=0.2.24",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.18",
    "pytest>=8.3.5",
    "pytest-dotenv>=0.5.2",
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.3",
    "dependency-injector>=4.47.1",
    "python-multipart>=0.0.20",
    "pymupdf>=1.26.0",
    "Pillow>=11.2.1",
    "openai>=1.84.0",
    "opensearch-py>=2.8.0",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "IPython>=9.3.0",
    "langchain-tavily>=0.2.11",
    "langchain-core>=0.3.65",
    "bs4>=0.0.2",
    "langchain-anthropic>=0.3.19",
    "streamlit>=1.49.1",
    "watchdog>=6.0.0",
    "pyinstaller>=6.15.0",
]
requires-python = ">=3.12,<3.15"
license = {text = "MIT"}

[tool.pdm]
distribution = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
maui_ui  = { cmd = "streamlit run src/maui_qa.py", env = { UI = "streamlit" } }
